+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## UNRELEASED
 
 ### Adds
+
+- Support for automatic inline style attribute sanitization for Rich Text widgets.
+- Adds text align controls for Rich Text widgets. The following tools are now supported as part of a rich text widget's `toolbar` property:
+-- `alignLeft`
+-- `alignRight`
+-- `alignCenter`
+-- `alignJustify`
 - Adds browser-side `editMode` flag that tracks the state of the current view (edit or preview), located at `window.apos.adminBar.editMode`
 - `@apostrophecms/express` module now supports the `trustProxy: true` option, allowing your reverse proxy server (such as nginx) to pass on the original hostname, protocol and client IP address.